+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,0 +1,9 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_content'];
+
+/**
+ * Fields
+ */
+$lang['newsPaginationTitle'][0] = 'Abschnittsüberschrift';
+$lang['newsPaginationTitle'][1] = 'Geben Sie hier eine Überschrift für die Paginierung ein.';